requirements-dev.txt
# 30x SEO - Development Dependencies

# Testing
pytest>=8.0.0,<9.0.0
pytest-cov>=5.0.0,<6.0.0

# Linting & Formatting
ruff>=0.9.0,<1.0.0

# YAML validation (used by validate_skills.py)
pyyaml>=6.0.0,<7.0.0
